pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-civic_desk"
version = "1.0.0"
description = "AI-Powered Civic Service Benchmarking Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.1",

    # RL Frameworks & Environments
    "gymnasium>=0.29.1",
    "stable-baselines3[extra]>=2.2.1",

    # Dashboard & Visualization
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "pandas>=2.0.0",

    # LLM Inference (benchmark & agent)
    "openai>=1.0.0",

    # HTTP client (agent.py)
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point
server = "civic_desk.server.app:main"
# Benchmark entry point
benchmark = "civic_desk.benchmark:run_benchmark"

[tool.setuptools]
include-package-data = true
packages = ["civic_desk", "civic_desk.server"]
package-dir = { "civic_desk" = ".", "civic_desk.server" = "server" }